% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{cdc_locator_set_credentials}
\alias{cdc_locator_set_credentials}
\title{Set CDC Locator API credentials}
\usage{
cdc_locator_set_credentials(client_id, client_token)
}
\arguments{
\item{client_id}{CDC client ID}

\item{client_token}{CDC client token}
}
\description{
Stores credentials in the user's .Renviron file. You only need to run
this once per machine. Restart R after setting credentials.
}
