% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{.cdc_locator_services}
\alias{.cdc_locator_services}
\title{Valid service enums for the CDC Locator API}
\format{
An object of class \code{character} of length 14.
}
\usage{
.cdc_locator_services
}
\description{
These values are passed to the \code{services} query parameter of the
\verb{/providers} endpoint as a comma-separated list.
}
\details{
The following enums are supported:
\itemize{
\item hivtesting       : HIV Testing
\item stdtesting       : STD Testing
\item selftesting      : Self-administered HIV & STD testing
\item preptesting     : Pre-Exposure Prophylaxis (PrEP)
\item pep              : Post-Exposure Prophylaxis (PEP)
\item condom           : Condom distribution
\item Hepc             : Hepatitis C testing
\item doxypep          : Doxycycline post-exposure prophylaxis
\item mailintesting    : Mail-in HIV self-testing kits
\item ryanwhite        : Ryan White HIV/AIDS Program services
\item clinics          : HRSA-funded health clinics
\item substanceabuse   : Substance use disorder treatment
\item mentalhealth     : Mental health services
\item housing          : Housing assistance
}
}
\keyword{internal}
