% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{.has_cdc_locator_credentials}
\alias{.has_cdc_locator_credentials}
\title{Check whether CDC Locator credentials are available}
\usage{
.has_cdc_locator_credentials()
}
\value{
Logical scalar
}
\description{
Check whether CDC Locator credentials are available
}
\keyword{internal}
