% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{.validate_cdc_locator_services}
\alias{.validate_cdc_locator_services}
\title{Validate CDC Locator service enums}
\usage{
.validate_cdc_locator_services(services)
}
\arguments{
\item{services}{Character vector of service enums.}
}
\value{
Invisibly returns TRUE if valid.
}
\description{
Validate CDC Locator service enums
}
\keyword{internal}
